Extract `DashCardActionButtons` component
