+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionButtons/DashCardActionButtons.jsx
@@ -1,0 +1,93 @@
+/* eslint-disable react/prop-types */
+import React from "react";
+import _ from "underscore";
+import { t } from "ttag";
+
+import Icon from "metabase/components/Icon";
+import Tooltip from "metabase/components/Tooltip";
+
+import { getVisualizationRaw } from "metabase/visualizations";
+
+import AddSeriesButton from "./AddSeriesButton";
+import ChartSettingsButton from "./ChartSettingsButton";
+import RemoveButton from "./RemoveButton";
+import ToggleCardPreviewButton from "./ToggleCardPreviewButton";
+
+import { HEADER_ACTION_STYLE } from "./constants";
+
+function DashCardActionButtons({
+  series,
+  isLoading,
+  isVirtualDashCard,
+  hasError,
+  onRemove,
+  onAddSeries,
+  onReplaceAllVisualizationSettings,
+  showClickBehaviorSidebar,
+  onPreviewToggle,
+  isPreviewing,
+  dashboard,
+}) {
+  const buttons = [];
+
+  if (getVisualizationRaw(series).visualization.supportPreviewing) {
+    buttons.push(
+      <ToggleCardPreviewButton
+        key="toggle-card-preview-button"
+        isPreviewing={isPreviewing}
+        onPreviewToggle={onPreviewToggle}
+      />,
+    );
+  }
+
+  if (!isLoading && !hasError) {
+    if (
+      onReplaceAllVisualizationSettings &&
+      !getVisualizationRaw(series).visualization.disableSettingsConfig
+    ) {
+      buttons.push(
+        <ChartSettingsButton
+          key="chart-settings-button"
+          series={series}
+          onReplaceAllVisualizationSettings={onReplaceAllVisualizationSettings}
+          dashboard={dashboard}
+        />,
+      );
+    }
+    if (!isVirtualDashCard) {
+      buttons.push(
+        <Tooltip key="click-behavior-tooltip" tooltip={t`Click behavior`}>
+          <a
+            className="text-dark-hover drag-disabled mr1"
+            data-metabase-event="Dashboard;Open Click Behavior Sidebar"
+            onClick={showClickBehaviorSidebar}
+            style={HEADER_ACTION_STYLE}
+          >
+            <Icon name="click" />
+          </a>
+        </Tooltip>,
+      );
+    }
+
+    if (getVisualizationRaw(series).visualization.supportsSeries) {
+      buttons.push(
+        <AddSeriesButton
+          key="add-series-button"
+          series={series}
+          onAddSeries={onAddSeries}
+        />,
+      );
+    }
+  }
+
+  return (
+    <span className="flex align-center text-medium" style={{ lineHeight: 1 }}>
+      {buttons}
+      <Tooltip tooltip={t`Remove`}>
+        <RemoveButton className="ml1" onRemove={onRemove} />
+      </Tooltip>
+    </span>
+  );
+}
+
+export default DashCardActionButtons;
