+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionButtons/index.ts
@@ -1,4 +1,1 @@
-export { default as AddSeriesButton } from "./AddSeriesButton";
-export { default as ChartSettingsButton } from "./ChartSettingsButton";
-export { default as RemoveButton } from "./RemoveButton";
-export { default as ToggleCardPreviewButton } from "./ToggleCardPreviewButton";
+export { default } from "./DashCardActionButtons";
